requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
openai-whisper @ git+https://github.com/openai/whisper.git
ffmpeg-python==0.2.0
numpy>=1.24
torch
soundfile==0.12.1